+++ changelog.md
@@ -9,6 +9,8 @@
 - Fixed placeholder on messages
 - Fixed the `/skull` command and added support for hexadecimal format
 - Fixed vault default slot by permission
+- Fixed placeholder on messages
+- Fixed duplicate key with power tools
 
 # 1.0.1.7